YoloCamera/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
pytesseract==0.3.10
ultralytics==8.0.196
Pillow==10.0.1
numpy==1.24.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0 